Create ASC758 driver 16.c
This right justified program uses breadth of ADC to read 2 current sensors alternating. Where 15 integers of range is thin for change in current (510-525), it's better than the (127-131) for a current domain of 0-1.6A. because motors are 4amps under no load, up to 20 amps sufficiently loaded, this program should operate 510-710 integer with no differential circuitry or change in reference voltage. Efforts to improve circuits are being investigated, Stu no longer developing this program.